Relocates DNS and resolve.conf parsing from network to core/network.
This serves the ultimate goal of removing the network package.
